basic client->server code working for positioning

+++ src/4ha6EW2cru.System/Platform/Win32Clock.h
@@ -23,7 +23,7 @@ public:
 
 	float GetDeltaMilliseconds( );
 
-	inline float GetTime( ) const { return timeGetTime( ); };
+	inline float GetTime( ) const { return static_cast< float >( timeGetTime( ) ); };
 
 private:
